add firebase connection
actualisation of gradle and import of google service

also deletion of test classes
