--- app/src/androidTest/java/com/example/unser_hoersaal/MainActivityTest.java
@@ -1,59 +1,0 @@
-package com.example.unser_hoersaal;
-import static org.junit.Assert.assertNotNull;
-
-import androidx.test.rule.ActivityTestRule;
-import androidx.test.runner.AndroidJUnit4;
-
-import androidx.test.rule.ActivityTestRule;
-
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import static androidx.test.espresso.Espresso.onView;
-import static androidx.test.espresso.action.ViewActions.click;
-import static androidx.test.espresso.action.ViewActions.closeSoftKeyboard;
-import static androidx.test.espresso.action.ViewActions.typeText;
-import static androidx.test.espresso.assertion.ViewAssertions.matches;
-
-import static androidx.test.espresso.matcher.ViewMatchers.withId;
-import static androidx.test.espresso.matcher.ViewMatchers.withText;
-
-public class MainActivityTest extends MainActivity {
-
-    @Rule
-    public ActivityTestRule<MainActivity> mActivityRule = new ActivityTestRule<>(MainActivity.class);
-
-    private MainActivity mTestActivity;
-
-    public MainActivityTest() {
-        mTestActivity = new MainActivity();
-        testPreconditions();
-
-    }
-    public void testPreconditions() {
-        assertNotNull("mTestActivity is null", mTestActivity);
-    }
-
-    @Test
-    public void ensureTextChangesWork() {
-        // Type text and then press the button.
-        onView(withId(R.id.inputField))
-                .perform(typeText("HELLO"), closeSoftKeyboard());
-        onView(withId(R.id.changeText)).perform(click());
-
-        // Check that the text was changed.
-        onView(withId(R.id.inputField)).check(matches(withText("Lalala")));
-    }
-
-    @Test
-    public void changeText_newActivity() {
-        // Type text and then press the button.
-        onView(withId(R.id.inputField)).perform(typeText("NewText"),
-                closeSoftKeyboard());
-        onView(withId(R.id.switchActivity)).perform(click());
-
-        // This view is in a different Activity, no need to tell Espresso.
-        onView(withId(R.id.resultView)).check(matches(withText("NewText")));
-    }
-}
