--- app/src/main/java/com/example/unser_hoersaal/MainActivity.java
@@ -1,35 +1,0 @@
-package com.example.unser_hoersaal;
-
-import androidx.appcompat.app.AppCompatActivity;
-import android.app.ActionBar;
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.EditText;
-
-public class MainActivity extends AppCompatActivity {
-
-    EditText editText;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        editText = (EditText) findViewById(R.id.inputField);
-    }
-
-    public void onClick(View view) {
-        switch (view.getId()) {
-            case R.id.changeText:
-                editText.setText("Lalala");
-                break;
-            case R.id.switchActivity:
-                Intent intent = new Intent(this, SecondActivity.class);
-                intent.putExtra("input", editText.getText().toString());
-                startActivity(intent);
-                break;
-        }
-
-    }
-}
