+++ app/src/main/java/views/EnterOrCreateCourseFragment.java
@@ -1,9 +1,10 @@
-package com.example.unser_hoersaal.views;
+package views;
 
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.cardview.widget.CardView;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
@@ -11,21 +12,15 @@ import androidx.navigation.Navigation;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.CheckBox;
-import android.widget.TextView;
 
 import com.example.unser_hoersaal.R;
 
+public class EnterOrCreateCourseFragment extends Fragment {
 
-public class LoginFragment extends Fragment {
+    CardView enterCourseCardView;
+    CardView createCourseCardView;
 
-    TextView registrationTextView;
-    TextView forgotPasswordTextView;
-    CheckBox keepLoggedInCheckBox;
-    Button loginButton;
-
-    public LoginFragment() {
+    public EnterOrCreateCourseFragment() {
         // Required empty public constructor
     }
 
@@ -38,7 +33,7 @@ public class LoginFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_login, container, false);
+        return inflater.inflate(R.layout.fragment_enter_or_create_course, container, false);
     }
 
     @Override
@@ -49,18 +44,14 @@ public class LoginFragment extends Fragment {
     }
 
     private void initUI(View view){
-        registrationTextView = view.findViewById(R.id.loginFragmentRegistrationTextView);
-        forgotPasswordTextView = view.findViewById(R.id.loginFragmentForgotPasswortTextView);
-        keepLoggedInCheckBox = view.findViewById(R.id.loginFragmentCheckBox);
-        loginButton = view.findViewById(R.id.loginFragmentLoginButton);
+        enterCourseCardView = view.findViewById(R.id.enterCourseCardView);
+        createCourseCardView = view.findViewById(R.id.createCourseCardView);
     }
 
     //setup Navigation to corresponding fragments
     private void setupNavigation(View view){
         NavController navController = Navigation.findNavController(view);
-
-        //todo add logic to login
-        registrationTextView.setOnClickListener(v -> navController.navigate(R.id.action_loginFragment_to_registrationFragment));
-        loginButton.setOnClickListener(v -> navController.navigate(R.id.action_loginFragment_to_enterOrCreateCourseFragment));
+        enterCourseCardView.setOnClickListener(v -> navController.navigate(R.id.action_enterOrCreateCourseFragment_to_enterCourseFragment));
+        createCourseCardView.setOnClickListener(v -> navController.navigate(R.id.action_enterOrCreateCourseFragment_to_createCourseFragment));
     }
 }
