+++ app/src/main/java/views/MainActivity.java
@@ -1,0 +1,22 @@
+package views;
+
+import androidx.appcompat.app.AppCompatActivity;
+import android.app.ActionBar;
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.EditText;
+
+import com.example.unser_hoersaal.R;
+
+
+public class MainActivity extends AppCompatActivity {
+
+    EditText editText;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+}
