+++ app/src/main/java/views/RegistrationFragment.java
@@ -1,4 +1,4 @@
-package com.example.unser_hoersaal.views;
+package views;
 
 import android.os.Bundle;
 
@@ -12,16 +12,20 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+import android.widget.CheckBox;
 import android.widget.EditText;
 
 import com.example.unser_hoersaal.R;
 
-public class EnterCourseFragment extends Fragment {
+public class RegistrationFragment extends Fragment {
 
-    EditText enterCourseEditText;
-    Button enterCourseButton;
+    EditText userNameEditText;
+    EditText passwordEditText;
+    EditText repeatPassowordEditText;
+    CheckBox checkBox;
+    Button registrationButton;
 
-    public EnterCourseFragment() {
+    public RegistrationFragment() {
         // Required empty public constructor
     }
 
@@ -34,7 +38,7 @@ public class EnterCourseFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_enter_course, container, false);
+        return inflater.inflate(R.layout.fragment_registration, container, false);
     }
 
     @Override
@@ -45,15 +49,18 @@ public class EnterCourseFragment extends Fragment {
     }
 
     private void initUI(View view){
-        enterCourseEditText = view.findViewById(R.id.enterCourseFragmentCourseNumberEditText);
-        enterCourseButton = view.findViewById(R.id.enterCourseFragmentEnterButton);
+        userNameEditText = view.findViewById(R.id.registrationFragmentUserNameEditText);
+        passwordEditText = view.findViewById(R.id.registrationFragmentPasswordEditText);
+        repeatPassowordEditText = view.findViewById(R.id.registrationFragmentRepeatPasswordEditText);
+        checkBox = view.findViewById(R.id.registrationFragmentCheckBox);
+        registrationButton = view.findViewById(R.id.registrationFragmentRegistrationButton);
     }
 
     //setup Navigation to corresponding fragments
     private void setupNavigation(View view){
         NavController navController = Navigation.findNavController(view);
 
-        //todo add logic to entering course
-        enterCourseButton.setOnClickListener(v -> navController.navigate(R.id.action_enterCourseFragment_to_currentCourseFragment));
+        //todo add logic to registration
+        registrationButton.setOnClickListener(v -> navController.navigate(R.id.action_registrationFragment_to_enterOrCreateCourseFragment));
     }
 }
