+++ app/src/main/res/layout/fragment_enter_course.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".views.EnterCourseFragment"
+    tools:context="views.EnterCourseFragment"
     android:background="@drawable/app_background">
 
     <androidx.constraintlayout.widget.ConstraintLayout
