+++ app/src/main/res/layout/fragment_enter_or_create_course.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/app_background"
-    tools:context=".views.EnterOrCreateCourseFragment">
+    tools:context="views.EnterOrCreateCourseFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
