+++ app/src/main/res/layout/fragment_login.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".views.LoginFragment"
+    tools:context="views.LoginFragment"
     android:orientation="vertical"
     android:background="@drawable/app_background">
