+++ build.gradle
@@ -6,6 +6,7 @@ buildscript {
     }
     dependencies {
         classpath "com.android.tools.build:gradle:7.0.4"
+        classpath "com.google.gms:google-services:4.3.10"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
