+++ google-services.json
@@ -1,0 +1,39 @@
+{
+  "project_info": {
+    "project_number": "397509465756",
+    "project_id": "unser-horsaal",
+    "storage_bucket": "unser-horsaal.appspot.com"
+  },
+  "client": [
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:397509465756:android:14987c76030c4d9e887f59",
+        "android_client_info": {
+          "package_name": "com.example.unser_hoersaal"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "397509465756-5vh08ml7jgk1ien46rrh8iesir28s2k0.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyDKpuIeUqrtO6HLR1MN4_EMrgQ9ej5AZUY"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "397509465756-5vh08ml7jgk1ien46rrh8iesir28s2k0.apps.googleusercontent.com",
+              "client_type": 3
+            }
+          ]
+        }
+      }
+    }
+  ],
+  "configuration_version": "1"
+}
